Added pakages: Command, parser, and FileIO, changed balance to BigDecimal, updated withdraw/deposit and tests

+++ src/main/java/com/revature/Account/AccountInterface.java
@@ -1,11 +1,13 @@
 package com.revature.Account;
 
+import java.math.BigDecimal;
+
 public interface AccountInterface {
     
     public int getAccountID();
     public void setAccountID(int id);
-    public int getBalance();
-    public void setBalance(int balance);
+    public BigDecimal getBalance();
+    public void setBalance(BigDecimal balance);
     public short getPIN();
     public void setPIN(short pin);
     public String getFirstName();
@@ -14,8 +16,8 @@ public interface AccountInterface {
     public void setLastName(String lastName);
     public String getAddress();
     public void setAddress(String address);
-    public void withdraw(int amount);
-    public void deposit(int amount);
+    public void withdraw(BigDecimal amount);
+    public void deposit(BigDecimal amount);
     public boolean getIsLocked();
     public void setIsLocked(boolean lock);
